% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEclear.R
\docType{data}
\name{BEclear example methylation data}
\alias{BEclear example methylation data}
\alias{ex.data}
\alias{BEclearData}
\title{Example data set for the BEclear-package}
\format{An example data matrix that is filled with beta values originally
stemming from breast cancer data from the TCGA portal [1], colnames are
sample ids, rownames are gene names. Generally, beta values are calculated by
dividing the methylated signal by the sum of the unmethylated and methylated
signals from a DNA methylation microrarray. The sample data used here
contains averaged beta values of probes that belong to promoter regions of
single genes. Another possibility would be to use beta values of single
probes, whereby the probe names should then be used instead of the gene names
as rownames of the matrix.}
\usage{
data(BEclearData)
}
\description{
Example data set for the BEclear-package
}
\references{
\insertRef{TCGA}{BEclear}
}
\keyword{datasets}
