% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEclear.R
\docType{data}
\name{BEclear example sample data}
\alias{BEclear example sample data}
\alias{ex.samples}
\title{Example data set for the BEclear-package}
\format{
An example data frame containing a column for the sample id and a
column for the corresponding batch id, stemming from breast cancer data
from the TCGA portal [1]
}
\usage{
data(BEclearData)
}
\description{
Example data set for the BEclear-package
}
\references{
\insertRef{TCGA}{BEclear}
}
\keyword{datasets}
