% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBatchEffectsForBatch.R
\name{calcBatchEffectsForBatch}
\alias{calcBatchEffectsForBatch}
\title{calcBatchEffectsForBatch}
\usage{
calcBatchEffectsForBatch(batch, samples, data, BPPARAM = bpparam())
}
\value{
the medians p-values for genes in a batch
}
\description{
function to calculate batch effects for every gene in a batch
}
\keyword{internal}
