% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBlockFrame.R
\name{calcBlockFrame}
\alias{calcBlockFrame}
\title{calcBlockFrame}
\usage{
calcBlockFrame(rowPositions, colPositions, rowBlockSize, colBlockSize)
}
\value{
a data.frame containing the start and stop positions for each block
}
\description{
calculates every start and stop position for every block
}
\keyword{internal}
