% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMediansForBatch.R
\name{calcMediansForBatch}
\alias{calcMediansForBatch}
\title{calcMediansForBatch}
\usage{
calcMediansForBatch(batch, genes, medianDif, samples, data)
}
\value{
the medians for genes in a batch
}
\description{
function to calculate medians for every gene for a batch
}
\keyword{internal}
