% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPositions.R
\name{calcPositions}
\alias{calcPositions}
\title{calcPositions}
\usage{
calcPositions(num, blockSize)
}
\value{
returns vector with number of blocks, start - and stop position of
last 2 blocks
}
\description{
calcPositions
}
\keyword{internal}
