% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPvalsForBatch.R
\name{calcPvalsForBatch}
\alias{calcPvalsForBatch}
\title{calcPvalsForBatch}
\usage{
calcPvalsForBatch(batch, samples, data, BPPARAM = bpparam())
}
\value{
the p-values for genes in a batch
}
\description{
function to calculate p-values for every gene in a batch
}
\keyword{internal}
