% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPvalsForGene.R
\name{calcPvalsForGene}
\alias{calcPvalsForGene}
\title{calcPvalsForGene}
\usage{
calcPvalsForGene(data)
}
\value{
the p-values for genes in a batch
}
\description{
function to calculate p-values for one gene in a batch
}
\keyword{internal}
