% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineBlocks.R
\name{combineBlocks}
\alias{combineBlocks}
\title{combineBlocks}
\usage{
combineBlocks(blockFrame, rowPos, colPos, dir)
}
\value{
the combined matrix consisting of the individual blocks
}
\description{
load and combine packages
}
\keyword{internal}
