% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlossp.R
\name{dlossp}
\alias{dlossp}
\title{dlossp}
\usage{
dlossp(L, R, lambda, x, nnzis, nnzjs, i, j, r)
}
\value{
the result as a list (elements can be accessed with x$Li or x$Ri, 
where x is the returned list)
}
\description{
compute the local loss of the p-th revealed entry for 
factorization LR
}
\keyword{internal}
