\name{ex.corrected.data}
\alias{ex.corrected.data}
\docType{data}
\title{Example matrix of corrected data for the BEclear-package}
\description{
Example matrix containing a already batch effect corrected sample matrix of
beta values from breast invasive carcinoma TCGA methylation data. The matrix
contains a small amount of wrongly predicted beta values to show the operating
principles of some of the methods from the BEclear package.
}
\usage{data(BEclearCorrected)}
\format{A matrix containing already corrected beta values of some samples from
the breast invasive carcinoma TCGA methylation data. The colnames denote
samples, rownames denote gene names.}
\value{A matrix (ex.corrected.data)}
\source{http://cancergenome.nih.gov/}
\keyword{datasets}
\keyword{ex.data}
\keyword{BEclear}