% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdepoch.R
\name{gdepoch}
\alias{gdepoch}
\title{gdepoch}
\usage{
gdepoch(L, R, lambda, eps, nnzis, nnzjs, is, js, D, m, n, r)
}
\value{
a list containing two matrices generated by the gradient descent for 
the current epoch
}
\description{
Run a gradient descent epoch (L and R are starting points, 
eps is stepsize)
}
\keyword{internal}
