% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMissingDataForBlock.R
\name{imputeMissingDataForBlock}
\alias{imputeMissingDataForBlock}
\title{imputeMissingDataForBlock}
\usage{
imputeMissingDataForBlock(block, dir, epochs, lambda = 1, gamma = 0.01, r = 10)
}
\value{
number of the block processed
}
\description{
imputeMissingDataForBlock
}
\keyword{internal}
