% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{localLoss}
\alias{localLoss}
\title{Calculating the Local Loss}
\usage{
localLoss(L, R, is, js, error_matrix)
}
\value{
a list containing two matrices containing the local loss
}
\description{
Calculates the gradient dR and dL for the current epoch.
}
\keyword{internal}
