% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnz.R
\name{nnz}
\alias{nnz}
\title{nnz}
\usage{
nnz(indexes, l = max(indexes))
}
\value{
a vector n containing the number of non-zero entries per index
}
\description{
counts for every index, the number of non-zero entries
}
\keyword{internal}
