% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGradientDescent.R
\name{runGradientDescent}
\alias{runGradientDescent}
\title{runGradientDescent}
\usage{
runGradientDescent(L0r10, R0r10, lambda, epochs, eps = 0.01, block,
  nnzis, nnzjs, is, js, D, m, n, r)
}
\value{
a list containing two matrices generated by the gradient descent
}
\description{
Runner for gradient descent (or stochastic gradient descent) for
the specified number of epoch
}
\keyword{internal}
