% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEclear.R
\docType{data}
\name{BEclear example sample data}
\alias{BEclear example sample data}
\alias{ex.samples}
\alias{BEclear}
\alias{example}
\alias{methylation}
\alias{data}
\title{Example data set for the BEclear-package}
\format{An example data frame containing a columnfor the sample id and a 
column for the corresponding batch id, stemming from breast cancer data 
from the TCGA portal [1]}
\usage{
data(BEclearData)
}
\description{
Example data set for the BEclear-package
}
\references{
[1] \url{http://cancergenome.nih.gov/}
}
\keyword{datasets}
