% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPvalues.R
\name{calcPvalues}
\alias{calcPvalues}
\title{Calculate p-values}
\usage{
calcPvalues(data, samples, adjusted=TRUE, method="fdr", 
BPPARAM=bpparam())
}
\arguments{
\item{data}{any matrix filled with beta values, column names have to be 
sample_ids corresponding to the ids listed in "samples", row names have to
be gene names.}

\item{samples}{data frame with two columns, the first column has to contain 
the sample numbers, the second column has to contain the corresponding batch
number. Colnames have to be named as "sample_id" and "batch_id".}

\item{adjusted}{should the p-values be adjusted or not, see "method" for 
available adjustment methods.}

\item{method}{adjustment method for p-value adjustment (if TRUE), default 
method is "false discovery rate adjustment", for other available methods 
see the description of the used standard R package \code{\link{p.adjust}}.}

\item{BPPARAM}{An instance of the 
\code{\link[BiocParallel]{BiocParallelParam}} that determines how to
parallelisation of the functions will be evaluated.}
}
\value{
a matrix containing p-values for all genes in all batches, the 
column names define the batch numbers, row names are the same gene names as 
contained in the input matrix.
}
\description{
Compares the distribution of all beta values corresponding to 
one batch with the distribution of all beta values corresponding to all 
other batches and retuns a p-value which defines if the distributions are 
the same or not. Standard two sided Kolmogorov-Smirnov test is used to 
calculate the (adjusted) p-values.
}
\details{
calcPvalues
}
\examples{
## Shortly running example. For a more realistic example that takes 
## some more time, run the same procedure with the full BEclearData
## dataset.

## Calculate fdr-adjusted p-values in non-parallel mode
data(BEclearData)
ex.data <- ex.data[31:90,7:26]
ex.samples <- ex.samples[7:26,] 
pvals <- calcPvalues(data=ex.data, samples=ex.samples,method="fdr")
}
\seealso{
\code{\link{ks.test}}

\code{\link{p.adjust}}

\code{\link{correctBatchEffect}}
}
