% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEclear.R
\docType{data}
\name{ex.corrected.data}
\alias{ex.corrected.data}
\title{Example matrix of corrected data for the BEclear-package}
\format{A matrix containing already corrected beta values of some samples from
the breast invasive carcinoma TCGA methylation data. The colnames denote
samples, rownames denote gene names.}
\usage{
data(BEclearCorrected)
}
\description{
Example matrix containing a already batch effect corrected sample
matrix of beta values from breast invasive carcinoma TCGA methylation
data.[1] The matrix contains a small amount of predicted beta values outside of
the boundaries to show the operating principles of some of the methods from 
the BEclear package.
}
\references{
\insertRef{TCGA}{BEclear}
}
\keyword{datasets}
