% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMissingDataForBlock.R
\name{imputeMissingDataForBlock}
\alias{imputeMissingDataForBlock}
\title{imputeMissingDataForBlock}
\usage{
imputeMissingDataForBlock(block, dir, epochs, lambda = 1, gamma = 0.01,
  r = 10, matrixOfOnes = FALSE)
}
\arguments{
\item{matrixOfOnes}{instead of starting with a random matrix, start from a matrix
of ones. For testing purposes only!}
}
\value{
number of the block processed
}
\description{
imputeMissingDataForBlock
}
\keyword{internal}
